+++ layouts/admin.vue
@@ -1,0 +1,15 @@
+<template>
+  <DefaultLayout>
+    <template #nav>
+      <AuthNav />
+    </template>
+    <Nuxt />
+  </DefaultLayout>
+</template>
+<script>
+  import DefaultLayout from './default.vue'
+  export default {
+    components: { DefaultLayout },
+    middleware: ['authenticated', 'admin'],
+  }
+</script>
