+++ layouts/default.vue
@@ -1,9 +1,13 @@
 <template>
   <v-app>
-    <GuestNav />
-    <v-container class="main-container" fluid>
-      <Nuxt />
-    </v-container>
+    <slot name="nav">
+      <AppNav />
+    </slot>
+    <v-main>
+      <v-container class="main-container" fluid>
+        <Nuxt />
+      </v-container>
+    </v-main>
     <ToastSnackbar />
   </v-app>
 </template>
