+++ layouts/nav.vue
@@ -11,15 +11,7 @@
 </template>
 
 <script>
-import AppHeader from '@/components/AppHeader.vue'
-import AppSideBar from '@/components/AppSideBar.vue'
 export default {
-  name: 'NavLayout',
-  components: { AppHeader, AppSideBar },
-  // comments: {
-  //   AppHeader,
-  //   AppSideBar
-  // },
   data() {
     return {}
   },
