+++ pages/administrator/_id/index.vue
@@ -1,3 +1,8 @@
 <template>
   <v-row> Admin page </v-row>
 </template>
+<script>
+  export default {
+    layout: 'admin',
+  }
+</script>
